Corrected test cases
