+++ client/src/test/java/io/pravega/client/watermark/WatermarkSerializerTest.java
@@ -29,7 +29,7 @@ public class WatermarkSerializerTest {
         SegmentWithRange segmentWithRange1 = new SegmentWithRange(0L, 0.0, 0.5);
         SegmentWithRange segmentWithRange2 = new SegmentWithRange(1L, 0.5, 1.0);
         ImmutableMap<SegmentWithRange, Long> map = ImmutableMap.of(segmentWithRange1, 1L, segmentWithRange2, 1L);
-        Watermark watermark = new Watermark(0L, 1L, map, "scope","stream");
+        Watermark watermark = new Watermark(0L, 1L, map, "scope", "stream");
         WatermarkSerializer serializer = new WatermarkSerializer();
         ByteBuffer serialized = serializer.serialize(watermark);
         Watermark deserialized = serializer.deserialize(serialized);
